chore(typescript): convert victory-brush-line to TS
